add Storage which deals with all file IOs
