+++ src/main/java/ssagit/data/taskList.txt
@@ -1,0 +1,6 @@
+todo | not done | b
+deadline | not done | return book  | 2/12/2019 1800
+deadline | not done | return book  | 2/12/2019 1800
+todo | not done | 1
+todo | not done | 2
+todo | not done | 3
