+++ src/main/java/ssagit/datevalidator/DateValidator.java
@@ -1,0 +1,8 @@
+package ssagit.datevalidator;
+
+/**
+ * Interface to validate String date input
+ */
+public interface DateValidator {
+    boolean isValid(String dateStr);
+}
